+++ sdk/databricks/azure-mgmt-databricks/azure/mgmt/databricks/aio/_client.py
@@ -1,4 +1,3 @@
-# pylint: disable=line-too-long,useless-suppression
 # coding=utf-8
 # --------------------------------------------------------------------------
 # Copyright (c) Microsoft Corporation. All rights reserved.
@@ -23,7 +22,6 @@ from ._configuration import AzureDatabricksManagementClientConfiguration
 from .operations import (
     AccessConnectorsOperations,
     Operations,
-    OutboundNetworkDependenciesEndpointsOperations,
     PrivateEndpointConnectionsOperations,
     PrivateLinkResourcesOperations,
     VNetPeeringOperations,
@@ -40,7 +38,7 @@ if TYPE_CHECKING:
     from azure.core.credentials_async import AsyncTokenCredential
 
 
-class AzureDatabricksManagementClient:  # pylint: disable=too-many-instance-attributes,docstring-keyword-should-match-keyword-only
+class AzureDatabricksManagementClient:  # pylint: disable=docstring-keyword-should-match-keyword-only
     """ARM Databricks.
 
     :ivar operations: Operations operations
@@ -52,10 +50,6 @@ class AzureDatabricksManagementClient:  # pylint: disable=too-many-instance-attr
      azure.mgmt.databricks.aio.operations.PrivateEndpointConnectionsOperations
     :ivar access_connectors: AccessConnectorsOperations operations
     :vartype access_connectors: azure.mgmt.databricks.aio.operations.AccessConnectorsOperations
-    :ivar outbound_network_dependencies_endpoints: OutboundNetworkDependenciesEndpointsOperations
-     operations
-    :vartype outbound_network_dependencies_endpoints:
-     azure.mgmt.databricks.aio.operations.OutboundNetworkDependenciesEndpointsOperations
     :ivar private_link_resources: PrivateLinkResourcesOperations operations
     :vartype private_link_resources:
      azure.mgmt.databricks.aio.operations.PrivateLinkResourcesOperations
@@ -133,9 +127,6 @@ class AzureDatabricksManagementClient:  # pylint: disable=too-many-instance-attr
             self._client, self._config, self._serialize, self._deserialize
         )
         self.access_connectors = AccessConnectorsOperations(
-            self._client, self._config, self._serialize, self._deserialize
-        )
-        self.outbound_network_dependencies_endpoints = OutboundNetworkDependenciesEndpointsOperations(
             self._client, self._config, self._serialize, self._deserialize
         )
         self.private_link_resources = PrivateLinkResourcesOperations(