+++ sdk/databricks/azure-mgmt-databricks/azure/mgmt/databricks/operations/_operations.py
@@ -528,34 +528,6 @@ def build_access_connectors_list_by_subscription_request(  # pylint: disable=nam
     return HttpRequest(method="GET", url=_url, params=_params, headers=_headers, **kwargs)
 
 
-def build_outbound_network_dependencies_endpoints_list_request(  # pylint: disable=name-too-long
-    resource_group_name: str, workspace_name: str, subscription_id: str, **kwargs: Any
-) -> HttpRequest:
-    _headers = case_insensitive_dict(kwargs.pop("headers", {}) or {})
-    _params = case_insensitive_dict(kwargs.pop("params", {}) or {})
-
-    api_version: str = kwargs.pop("api_version", _params.pop("api-version", "2026-01-01"))
-    accept = _headers.pop("Accept", "application/json")
-
-    # Construct URL
-    _url = "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Databricks/workspaces/{workspaceName}/outboundNetworkDependenciesEndpoints"
-    path_format_arguments = {
-        "subscriptionId": _SERIALIZER.url("subscription_id", subscription_id, "str"),
-        "resourceGroupName": _SERIALIZER.url("resource_group_name", resource_group_name, "str"),
-        "workspaceName": _SERIALIZER.url("workspace_name", workspace_name, "str"),
-    }
-
-    _url: str = _url.format(**path_format_arguments)  # type: ignore
-
-    # Construct parameters
-    _params["api-version"] = _SERIALIZER.query("api_version", api_version, "str")
-
-    # Construct headers
-    _headers["Accept"] = _SERIALIZER.header("accept", accept, "str")
-
-    return HttpRequest(method="GET", url=_url, params=_params, headers=_headers, **kwargs)
-
-
 def build_private_link_resources_get_request(
     resource_group_name: str, workspace_name: str, group_id: str, subscription_id: str, **kwargs: Any
 ) -> HttpRequest:
@@ -3172,104 +3144,6 @@ class AccessConnectorsOperations:  # pylint: disable=docstring-missing-param
             return pipeline_response
 
         return ItemPaged(get_next, extract_data)
-
-
-class OutboundNetworkDependenciesEndpointsOperations:  # pylint: disable=docstring-missing-param,name-too-long
-    """
-    .. warning::
-        **DO NOT** instantiate this class directly.
-
-        Instead, you should access the following operations through
-        :class:`~azure.mgmt.databricks.AzureDatabricksManagementClient`'s
-        :attr:`outbound_network_dependencies_endpoints` attribute.
-    """
-
-    def __init__(self, *args, **kwargs) -> None:
-        input_args = list(args)
-        self._client: PipelineClient = input_args.pop(0) if input_args else kwargs.pop("client")
-        self._config: AzureDatabricksManagementClientConfiguration = (
-            input_args.pop(0) if input_args else kwargs.pop("config")
-        )
-        self._serialize: Serializer = input_args.pop(0) if input_args else kwargs.pop("serializer")
-        self._deserialize: Deserializer = input_args.pop(0) if input_args else kwargs.pop("deserializer")
-
-    @distributed_trace
-    def list(
-        self, resource_group_name: str, workspace_name: str, **kwargs: Any
-    ) -> List[_models.OutboundEnvironmentEndpoint]:
-        """Gets a list of egress endpoints (network endpoints of all outbound dependencies) in the
-        specified Workspace.
-
-        Gets the list of endpoints that VNET Injected Workspace calls Azure Databricks Control Plane.
-        You must configure outbound access with these endpoints. For more information, see
-        `https://docs.microsoft.com/en-us/azure/databricks/administration-guide/cloud-configurations/azure/udr
-        <https://docs.microsoft.com/en-us/azure/databricks/administration-guide/cloud-configurations/azure/udr>`_.
-
-        :param resource_group_name: The name of the resource group. The name is case insensitive.
-         Required.
-        :type resource_group_name: str
-        :param workspace_name: The name of the workspace. Required.
-        :type workspace_name: str
-        :return: list of OutboundEnvironmentEndpoint
-        :rtype: list[~azure.mgmt.databricks.models.OutboundEnvironmentEndpoint]
-        :raises ~azure.core.exceptions.HttpResponseError:
-        """
-        error_map: MutableMapping = {
-            401: ClientAuthenticationError,
-            404: ResourceNotFoundError,
-            409: ResourceExistsError,
-            304: ResourceNotModifiedError,
-        }
-        error_map.update(kwargs.pop("error_map", {}) or {})
-
-        _headers = kwargs.pop("headers", {}) or {}
-        _params = kwargs.pop("params", {}) or {}
-
-        cls: ClsType[List[_models.OutboundEnvironmentEndpoint]] = kwargs.pop("cls", None)
-
-        _request = build_outbound_network_dependencies_endpoints_list_request(
-            resource_group_name=resource_group_name,
-            workspace_name=workspace_name,
-            subscription_id=self._config.subscription_id,
-            api_version=self._config.api_version,
-            headers=_headers,
-            params=_params,
-        )
-        path_format_arguments = {
-            "endpoint": self._serialize.url("self._config.base_url", self._config.base_url, "str", skip_quote=True),
-        }
-        _request.url = self._client.format_url(_request.url, **path_format_arguments)
-
-        _decompress = kwargs.pop("decompress", True)
-        _stream = kwargs.pop("stream", False)
-        pipeline_response: PipelineResponse = self._client._pipeline.run(  # pylint: disable=protected-access
-            _request, stream=_stream, **kwargs
-        )
-
-        response = pipeline_response.http_response
-
-        if response.status_code not in [200]:
-            if _stream:
-                try:
-                    response.read()  # Load the body in memory and close the socket
-                except (StreamConsumedError, StreamClosedError):
-                    pass
-            map_error(status_code=response.status_code, response=response, error_map=error_map)
-            error = _failsafe_deserialize(
-                _models.ErrorResponse,
-                response,
-            )
-            raise HttpResponseError(response=response, model=error, error_format=ARMErrorFormat)
-
-        if _stream:
-            deserialized = response.iter_bytes() if _decompress else response.iter_raw()
-        else:
-            deserialized = _deserialize(List[_models.OutboundEnvironmentEndpoint], response.json())
-
-        if cls:
-            return cls(pipeline_response, deserialized, {})  # type: ignore
-
-        return deserialized  # type: ignore
 
 
 class PrivateLinkResourcesOperations:  # pylint: disable=docstring-missing-param